back-inference/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
Pillow
python-multipart
